feat: succession of different modals based on the number of clicks
