+++ src/tab/triggers.js
@@ -1,13 +1,14 @@
 import {detectPageLoad, domainExportChecker, domainLoadChecker} from "./checker/domainChecker";
 import {launchExport} from "./exportProcessor/exportProcess";
 import {uiEnhancer} from "./uiEnhancer/uiEnhancer";
-import ModalDetectClicks from "./uiEnhancer/modals/ModalDetectClicks";
+// import ModalDetectClicks from "./uiEnhancer/modals/ModalDetectClicks";
+import {checkClickCountAndDisplayModal} from "./uiEnhancer/modals/clickCount";
 
 export async function actionExtensionIconClicked() {
   const domainPage = await domainExportChecker();
   if (domainPage === null) return;
   launchExport(domainPage);
-  new ModalDetectClicks(domainPage).appendModal();
+  checkClickCountAndDisplayModal(domainPage);
 }
 
 export async function actionPageLoaded() {
